good in DX12

+++ truck-rendimpl/src/shaders/face.wgsl
@@ -2,7 +2,7 @@ struct VertexInput {
     [[location(0)]] position: vec3<f32>;
     [[location(1)]] uv: vec2<f32>;
     [[location(2)]] normal: vec3<f32>; 
-    [[location(3)]] boundary_range: vec2<f32>;
+    [[location(3)]] boundary_range: vec2<u32>;
 };
 
 [[block]]
@@ -66,7 +66,7 @@ struct VertexOutput {
     [[location(0)]] position: vec3<f32>;
     [[location(1)]] uv: vec2<f32>;
     [[location(2)]] normal: vec3<f32>; 
-    [[location(3)]] boundary_range: vec2<f32>;
+    [[location(3)]] boundary_range: vec2<u32>;
 };
 
 [[stage(vertex)]]
